requirements.txt
Django==4.0.6
djangorestframework==3.13.1
slack-sdk==3.17.2
mysqlclient==2.1.1
gunicorn==20.1.0